+++ kgcnn/hyper/hyper.py
@@ -1,7 +1,0 @@
-from kgcnn.training.hyper import *
-
-import logging
-
-# For old reference.
-logging.error(
-    "Module 'kgcnn.hyper' is deprecated and will be removed in future versions. Please move to 'kgcnn.training.hyper'.")