+++ kgcnn/mol/__init__.py
@@ -1,5 +1,0 @@
-import logging
-
-# For old reference.
-logging.error(
-    "Module 'kgcnn.mol' is deprecated and will be removed in future versions. Please move to 'kgcnn.molecule'.")