+++ kgcnn/mol/convert.py
@@ -1,1 +1,0 @@
-from kgcnn.molecule.convert import *