+++ kgcnn/mol/graph_babel.py
@@ -1,1 +1,0 @@
-from kgcnn.molecule.graph_babel import *