+++ kgcnn/mol/graph_rdkit.py
@@ -1,1 +1,0 @@
-from kgcnn.molecule.graph_rdkit import *