+++ kgcnn/moldyn/ase_calc.py
@@ -1,1 +1,0 @@
-from kgcnn.molecule.dynamics.ase_calc import *