+++ kgcnn/moldyn/base.py
@@ -1,1 +1,0 @@
-from kgcnn.molecule.dynamics.base import *